Cleanup: Move buildWithinViewportGqlGeometryFilter into Map
