+++ ui/src/components/map/queries/useGetMapTerminals.ts
@@ -6,9 +6,9 @@ import {
 } from '../../../generated/graphql';
 import { Operation } from '../../../redux';
 import { Viewport } from '../../../redux/types';
-import { buildWithinViewportGqlGeometryFilter } from '../../../utils';
 import { useMapDataLayerSimpleQueryLoader } from '../../common/hooks/useLoader';
 import { ChildStopPlaceIds, MapTerminal } from '../types';
+import { buildWithinViewportGqlGeometryFilter } from '../utils/buildWithinViewportGqlGeometryFilter';
 
 const GQL_QUERY_GET_STOP_TERMINALS_BY_LOCATION = gql`
   query GetStopTerminalsByLocation($locationFilter: geometry_comparison_exp) {
